backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.16.0",
    "mysqlclient>=2.2.7",
    "django-cors-headers>=4.3.1",
]
